% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnames.R
\name{addnames}
\alias{addnames}
\title{Add names by ID number to a Datalink Connect output .xlsx file}
\usage{
addnames(input, id)
}
\arguments{
\item{input}{.xlsx exported from Datalink Connect read into R by using the command: read.xlsx2(file,sheetName="Results Grid")}

\item{id}{a dataframe containing IDs and Names of students appearing in input file.  Column with ID numbers should be named "ID" and column with names should be named "Name"}
}
\value{
A the same dataframe read in as object with names added to the appropriate column
}
\description{
uses the .xlsx output from Datalink Connect software and will score responses by assigning a grade value to each possible response
}
