% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/both_ok.R
\name{both.ok}
\alias{both.ok}
\title{Will allow either marked or unmarked answers for a particular question and lead will be accepted.}
\usage{
both.ok(lead, grades, exam, answer.value = 0.6)
}
\arguments{
\item{lead}{The multiple choice question and answer for which either marked or unmarked would be acceptable.  This should be a character string with the question number immediately followed by the lead. For example, if either answer (marked or unmarked) is acceptable for lead B of question 3, the string should be "3B", or the argument should be lead = "3B". This function only works for one lead.}

\item{grades}{Output grades dataframe from grade.earned() or grade.missed() functions}

\item{exam}{Answer matrix from DataLink Excel format export}

\item{answer.value}{a single value or a vector equal to number of answers assigning values to each possible response}
}
\value{
grades dataframe is updated
}
\description{
Uses the exam answers read in from DataLink software, and import using read.xlsx function, and the grades output from grade.earned() or grade.missed() functions.
}
