% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{diagnostic}
\alias{diagnostic}
\title{Runs two diagnostics on multiple choice leads from multiple choice data}
\usage{
diagnostic(input)
}
\arguments{
\item{input}{.xlsx exported from Datalink Connect read into R by using the command: read.xlsx2(file,sheetName="Results Grid")}
}
\value{
An .xlsx spread sheet is written into the working directory
}
\description{
This function runs two diagnostics, difficulty index and discrimination index, on each lead of a multiple choice question matrix produced by a Datalink Connect software
}
