% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.earned.R
\name{grade.earned}
\alias{grade.earned}
\title{Grades output from Datalink and reports number of earned points}
\usage{
grade.earned(input, answer.value = 0.2, possible)
}
\arguments{
\item{input}{.xlsx exported from Datalink Connect read into R by using the command: read.xlsx2(file,sheetName="Results Grid")}

\item{answer.value}{a single value or a vector equal to number of answers assigning values to each possible response}

\item{possible}{a single value of total possible points}
}
\value{
A dataframe consisting of student names, ID numbers and number of points missed on scanned questions
}
\description{
uses the .xlsx output from Datalink Connect software and will score responses by assigning a grade value to each possible response
}
